Updates Auth flow (#1334)
- Invalid tokens in the URL no longer cause login to display empty name field.
- Adds automatic retry with local token if URL token is invalid (bookmarked links).

+++ charactersheet/charactersheet/services/common/account/authentication_service.js
@@ -24,6 +24,13 @@ function _AuthenticationService(config) {
      */
     self.config = config;
 
+    self._tokenOrigin = null;
+
+    self.TOKEN_ORIGINS = {
+        FRAGMENT: 'fragment',
+        LOCAL: 'local'
+    };
+
     self.init = function() {
         // Make sure the token we have is valid.
         var fragments = (new URI()).fragment(true);
@@ -32,37 +39,56 @@ function _AuthenticationService(config) {
         var accessToken = null;
         if (fragments['access_token']) {
             accessToken = fragments['access_token'];
+            self._tokenOrigin = self.TOKEN_ORIGINS.FRAGMENT;
         } else if (token && token.isValid()) {
             accessToken = token.accessToken();
+            self._tokenOrigin = self.TOKEN_ORIGINS.LOCAL;
         } else {
             return;
         }
 
-        Utility.oauth.getJSON(self.validationUrl, self._handleValidationResponse,
-            self._handleValidationResponse, accessToken);
+        self._doAuthCheck(accessToken);
     };
 
     // Private Methods
 
+    self._doAuthCheck = function(accessToken) {
+        Utility.oauth.getJSON(self.validationUrl, self._handleValidationResponse,
+            self._handleValidationResponse, accessToken);
+    };
+
     self._handleValidationResponse = function(data, status) {
-        var token = PersistenceService.findAll(AuthenticationToken)[0];
         if (status !== 'success' || data.code !== 0) {
-            if (token) {
-                token.delete();
+            var token = PersistenceService.findAll(AuthenticationToken)[0];
+            if (self._tokenOrigin == self.TOKEN_ORIGINS.FRAGMENT) {
+                // Retry with local token.
+                self._tokenOrigin = self.TOKEN_ORIGINS.LOCAL;
+                if (token) {
+                    self._doAuthCheck(token.accessToken());
+                }
+            } else {
+                // All tokens are invalid and should be removed.
+                if (token) {
+                    token.delete();
+                }
             }
-        } else {
-            var fragments = (new URI()).fragment(true);
+            return;
+        }
+
+        if (self._tokenOrigin == self.TOKEN_ORIGINS.FRAGMENT) {
+            var token = PersistenceService.findAll(AuthenticationToken)[0];
             if (!token) {
                 token = new AuthenticationToken();
             }
-
+            var fragments = (new URI()).fragment(true);
             token.mapTokenKeys(fragments);
             token.startTime((new Date()).getTime());
             token.save();
-
-            if (token.isValid()) {
-                Notifications.authentication.loggedIn.dispatch();
-            }
+            Notifications.authentication.loggedIn.dispatch();
+        } else if (self._tokenOrigin = self.TOKEN_ORIGINS.LOCAL) {
+            Notifications.authentication.loggedIn.dispatch();
+        } else {
+            return;
         }
     };
 }
